refactor: 리뷰 적용 `BusinessException` 서비스에서 발생할 수 있게 수정
기존 테스트 및 서비스 변경점
- 매개변수 변경
