+++ src/main/java/org/programmers/signalbuddyfinal/domain/trafficSignal/service/TrafficCsvService.java
@@ -15,15 +15,13 @@ import org.springframework.transaction.annotation.Transactional;
 import java.io.*;
 import java.nio.charset.Charset;
 
-import java.util.regex.Pattern;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.Reader;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.regex.Pattern;
 
 @Slf4j
 @Service
@@ -34,35 +32,49 @@ public class TrafficCsvService {
     private final TrafficRepository trafficRepository;
 
     @Transactional
-    public void saveCsvData(File file) throws IOException {
+    public void saveCsvData(String fileName) throws IOException {
+
+        File file = new File("src/main/resources/static/file/"+fileName);
 
         try (Reader reader = new BufferedReader(new InputStreamReader(new FileInputStream(file), Charset.forName("EUC-KR")))) {
 
-            try {
 
-                List<TrafficSignal> entityList = new ArrayList<>();
-
-                CsvToBean<TrafficFileResponse> csvToBean = new CsvToBeanBuilder<TrafficFileResponse>(reader)
-                        .withType(TrafficFileResponse.class)
-                        .withIgnoreLeadingWhiteSpace(true)
-                        .build();
-
-                List<TrafficFileResponse> traffics = csvToBean.parse();
-
-                for (TrafficFileResponse trafficRes : traffics) {
-                    entityList.add(new TrafficSignal(trafficRes));
-                }
-
-                trafficRepository.saveAll(entityList);
-
-            } catch (DataIntegrityViolationException e) {
-                log.error("❌ Data Integrity Violation: {}", e.getMessage(), e);
-                throw new BusinessException(TrafficErrorCode.ALREADY_EXIST_TRAFFIC_SIGNAL);
-            } catch (Exception e) {
-                log.error(e.getMessage());
+            if (!isValidFileName(fileName)) {
+                throw new SecurityException("경로 탐색 시도 감지됨");
             }
 
+            List<TrafficSignal> entityList = new ArrayList<>();
+
+            CsvToBean<TrafficFileResponse> csvToBean = new CsvToBeanBuilder<TrafficFileResponse>(reader)
+                    .withType(TrafficFileResponse.class)
+                    .withIgnoreLeadingWhiteSpace(true)
+                    .build();
+
+            List<TrafficFileResponse> traffics = csvToBean.parse();
+
+            for (TrafficFileResponse trafficRes : traffics) {
+                entityList.add(new TrafficSignal(trafficRes));
+            }
+
+            trafficRepository.saveAll(entityList);
+
+        } catch (FileNotFoundException e){
+            log.error("❌ File Not Found : {}", e.getMessage(), e);
+            throw new BusinessException(TrafficErrorCode.FILE_NOT_FOUND);
+        } catch (DataIntegrityViolationException e) {
+            log.error("❌ Data Integrity Violation: {}", e.getMessage(), e);
+            throw new BusinessException(TrafficErrorCode.ALREADY_EXIST_TRAFFIC_SIGNAL);
+        } catch (Exception e) {
+            log.error(e.getMessage());
         }
+
+    }
+
+    // 파일 이름 검증 (특수 문자 및 경로 탐색 방지)
+    private boolean isValidFileName(String fileName) {
+        String regex = "^[a-zA-Z0-9._-]+$";
+        Pattern pattern = Pattern.compile(regex);
+        return pattern.matcher(fileName).matches();
     }
 
 }
